Fix daemon self-monitor issue
Daemons were tailing the log file too early; when the server starts it moves the log to a backup which we continued tailing. Now we wait to see log activity in the designated log before tailing it.

Fixed an issue where a failed server monitor would break the status page